Fixed bug (changed vector type np.float64->np.float32 to fit into faiss)

+++ run.py
@@ -200,7 +200,7 @@ if args.inference is not None:
         if not os.path.exists(prefetch_config['nn_base_path']):
             print('''NN base doesn't exist for the model, build NN base...''')
             t0 = time.time()
-            vectors = np.load(prefetch_config['vectors_path'])
+            vectors = np.load(prefetch_config['vectors_path']).astype(np.float32)
             vector_idx = np.load(prefetch_config['vector_index_path'])
             nn_per_query = prefetch_config['n_prefetch']
             nn_base = prefetch.build_nn_base(gru.itemidmap, vectors, vector_idx, nn_per_query, item_key='ItemId')
